requirements.txt
factory_boy
pytest
pytest-cov
pytest-capturelog
vcrpy
